Support 페이지의 CSS를 수정한다
