+++ src/pages/support.jsx
@@ -4,6 +4,7 @@ import styled from "@emotion/styled";
 
 import PostTitle from "../components/posts/PostTitle";
 import ParticleNetwork from "../components/support/ParticleNetwork";
+import RespectSatoshi from "../components/support/RespectSatoshi";
 
 import LightningContainer from "../container/LightningContainer";
 import LayoutContainer from "../container/LayoutContainer";
@@ -15,8 +16,9 @@ const SupportSection = styled.section({
   minHeight: base.pageMinHeight,
   padding: `${unit(4)} ${unit(2)}`,
   display: "flex",
+  flexDirection: "column",
   alignItems: "center",
-  justifyContent: "center",
+  justifyContent: "space-between",
   textAlign: "center",
 });
 
@@ -26,10 +28,9 @@ export default function SupportPage({ location }) {
       <SeoContainer title="Support" pathname={location.pathname} />
       <ParticleNetwork />
       <SupportSection>
-        <div>
-          <PostTitle>Buy me a 막걸리🌾</PostTitle>
-          <LightningContainer />
-        </div>
+        <PostTitle>Buy me a 막걸리🌾</PostTitle>
+        <LightningContainer />
+        <RespectSatoshi />
       </SupportSection>
     </LayoutContainer>
   );
