+++ src/pages/support.test.jsx
@@ -73,6 +73,17 @@ describe("SupportPage", () => {
     expect(particleEl).toBeInTheDocument();
   });
 
+  it("사토시 나카모토에게 경의를 표하는 문구를 출력한다", () => {
+    const respectsEl = screen.getByText("Pay my respects to");
+    const satoshiEl = screen.getByText("Satoshi Nakamoto");
+
+    expect(respectsEl).toBeInTheDocument();
+    expect(satoshiEl.closest("a")).toHaveAttribute(
+      "href",
+      "https://bitcoin.org/bitcoin.pdf"
+    );
+  });
+
   it("footer를 출력한다", () => {
     const year = new Date().getFullYear();
     const copyrightEl = screen.getByText(
